# Tap2Pay/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OrchestratePay Backend API
  version: 1.0.0
  description: |
    NFC tap-to-pay backend for Kenyan merchants.
    Merchants authenticate with JWT. The Android app polls for payment status.
    M-Pesa callbacks arrive asynchronously from Safaricom's servers.
  contact:
    name: OrchestratePay
    url: https://orchestratepay.co.ke
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.orchestratepay.co.ke/api/v1
    description: Production

# ─── Security ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 8-hour JWT issued at login. Stored in Android EncryptedSharedPreferences.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Merchant not found or inactive
        details:
          type: string
          description: Only present on 400 validation errors
    LoginRequest:
      type: object
      required: [email, password, deviceId]
      properties:
        email:
          type: string
          format: email
          example: merchant@example.co.ke
        password:
          type: string
          minLength: 8
          example: S3cur3P@ss!
        deviceId:
          type: string
          description: Android device identifier for single-device session enforcement
          example: android-imei-867123456789012
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token, valid for 8 hours
        merchantId:
          type: string
          format: uuid
        merchantName:
          type: string
          example: Mama Ngina's Chemist
        expiresAt:
          type: integer
          description: Unix timestamp (ms) when the token expires
...
paths:
...
  /auth/login:
    post:
      summary: Merchant login
      description: |
        Issues an 8-hour JWT. Calling this from a new device invalidates the
        previous session on the old device (single-device enforcement).
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: Merchant logout
      description: Clears device binding, invalidating the current session.
      tags: [Auth]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...